feat(lib): added global argument handler
