+++ lib/core/console/argument_handler.rb
@@ -1,0 +1,28 @@
+module SecureKeys
+  module Core
+    module Console
+      class ArgumentHandler
+        class << self
+          attr_reader :arguments
+        end
+
+        # Configure the default arguments
+        @arguments = {
+          delimiter: nil,
+          identifier: nil,
+        }
+
+        # Fetch the argument value by key
+        # from CLI arguments or environment variables
+        #
+        # @param key [Symbol] the argument key
+        # @param default [String] the default value
+        #
+        # @return [String] the argument value
+        def self.fetch(key:, default: nil)
+          @arguments[key.to_sym] || ENV.fetch("secure_keys_#{key}".upcase, nil) || default
+        end
+      end
+    end
+  end
+end
